updated folders order, making mobile menu
everything is without images adaptation
