--- src/components/AuthNav/AuthNavContainer/AuthNavContainer.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-const AuthNavContainerM = styled.div``;
-
-export default AuthNavContainerM;
