--- src/components/AuthNav/AuthNavContainer/AuthNavContainer.jsx
@@ -1,7 +1,0 @@
-import AuthNavContainterM from './NavContainer/NavContainer';
-
-const AuthNavContainer = () => {
-  return <AuthNavContainterM></AuthNavContainterM>;
-};
-
-export default AuthNavContainer;
