--- src/components/Header/ButtonText/ButtonText.jsx
@@ -1,7 +1,0 @@
-import MainButtonText from "./ButtonText";
-
-const ButtonText = () => {
-    return <MainButtonText ></MainButtonText>;
-}
-
-export default ButtonText;
