+++ src/components/Header/Header.styled.js
@@ -1,23 +1,29 @@
 import styled from 'styled-components';
 
-const MainHeader = styled.header`
+export const MainHeader = styled.header`
   height: 28px;
   margin-top: 24px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
 
   @media screen and (min-width: 768px) {
     height: 44px;
     margin-top: 24px;
-    display: flex;
-    align-items: center;
   }
 
   @media screen and (min-width: 1280px) {
     height: 48px;
     margin-top: 20px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
   }
 `;
 
-export default MainHeader;
+export const Container = styled.div`
+  display: flex;
+`;
+
+export const UserNavContainer = styled.div`
+  display: flex;
+`;
+
+
