--- src/components/Header/Icon.jsx
@@ -1,34 +1,0 @@
-const Paw = () =>{return (
-  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-    <g clipPath="url(#clip0_40956_131)">
-      <path
-        d="M15.3613 14.3621C15.0946 12.3788 13.8528 10.6897 12.0393 9.84408C10.2259 8.99851 8.13364 9.13284 6.44302 10.2034L2.36887 12.7829C1.69287 13.211 1.18351 13.8319 0.895812 14.5784C0.608038 15.3251 0.569171 16.1272 0.783156 16.8984C0.997551 17.6695 1.44487 18.3367 2.07661 18.8281C2.70843 19.3193 3.46541 19.5884 4.26566 19.6062L4.31795 19.6072C6.32515 19.6661 8.23238 20.5633 9.55763 22.0739C9.89947 22.4634 10.3104 22.7714 10.7701 22.9858C11.02 23.1023 11.2847 23.1912 11.5602 23.2506C12.3426 23.4188 13.1413 23.3326 13.87 23.0013C14.5983 22.6697 15.1881 22.1243 15.5753 21.424C15.9625 20.7235 16.1108 19.9341 16.0042 19.141L15.3613 14.3621ZM13.2874 21.7207C12.3605 22.1422 11.2867 21.9114 10.6152 21.1461C9.81708 20.2364 8.84232 19.5131 7.76431 19.0104C6.68729 18.5082 5.50694 18.2263 4.29792 18.1995L4.2635 18.1988C3.25999 18.1624 2.40849 17.492 2.1386 16.5219C1.86646 15.541 2.26132 14.5161 3.1214 13.9715L7.19572 11.392C8.47945 10.5792 10.0677 10.4772 11.4446 11.1192C12.8215 11.7613 13.7645 13.0437 13.9671 14.5496L14.6099 19.3284C14.7455 20.3378 14.214 21.2991 13.2874 21.7207Z"
-        fill="#FEF9F9"
-      />
-      <path
-        d="M4.72141 9.39524C5.54872 9.45292 6.34539 9.13609 6.96485 8.50306C7.55409 7.90059 7.91268 7.07743 7.9743 6.18489C8.03553 5.29277 7.79336 4.42827 7.29259 3.75049C6.76615 3.03851 6.02075 2.61496 5.19438 2.55832C3.53247 2.44301 2.0737 3.88292 1.94231 5.76906C1.8395 7.27113 2.60962 8.6087 3.76712 9.14845C4.06268 9.28627 4.38352 9.37204 4.72141 9.39524ZM3.34575 5.86588C3.42319 4.75444 4.20887 3.90011 5.09756 3.96176C5.25809 3.97278 5.41412 4.01382 5.56098 4.0823C5.78767 4.18801 5.99296 4.35909 6.16113 4.58682C6.4637 4.996 6.60917 5.52933 6.57079 6.08824C6.53208 6.6474 6.31495 7.15549 5.95913 7.5191C5.63302 7.85232 5.22802 8.02026 4.81905 7.99178L4.81823 7.99181C3.93038 7.93095 3.26986 6.97704 3.34575 5.86588Z"
-        fill="#FEF9F9"
-      />
-      <path
-        d="M10.6413 8.64674C12.403 9.46825 14.5978 8.50366 15.5337 6.4965C16.4699 4.48884 15.7982 2.1872 14.0365 1.36569C12.2749 0.544262 10.0802 1.50926 9.144 3.51692C8.20804 5.52408 8.87973 7.82531 10.6413 8.64674ZM13.442 2.64068C14.5006 3.13431 14.8671 4.59743 14.2587 5.90196C13.6507 7.206 12.2944 7.86538 11.2358 7.37175C10.1772 6.87812 9.81091 5.41549 10.419 4.11146C11.0273 2.80692 12.3834 2.14705 13.442 2.64068Z"
-        fill="#FEF9F9"
-      />
-      <path
-        d="M15.7437 12.1141L15.7439 12.1142C15.973 12.3305 16.2294 12.5024 16.5047 12.6308C17.7895 13.2299 19.4776 12.8786 20.6256 11.6643C21.2865 10.9654 21.6763 10.0822 21.7231 9.17769C21.7722 8.23102 21.4445 7.37435 20.8004 6.76527C20.1561 6.15653 19.2819 5.8774 18.3394 5.97975C17.4389 6.07767 16.5793 6.51648 15.9189 7.21567C14.5251 8.69034 14.4466 10.8878 15.7437 12.1141ZM16.9413 8.18197C17.3726 7.72545 17.923 7.44025 18.4915 7.37844C19.0174 7.32122 19.4943 7.46662 19.8341 7.78771C20.1736 8.10889 20.3455 8.57654 20.3181 9.10483C20.2885 9.67571 20.0348 10.2415 19.6032 10.6978C18.7422 11.6088 17.4444 11.7855 16.7102 11.0918C15.9768 10.3982 16.0804 9.09291 16.9413 8.18197Z"
-        fill="#FEF9F9"
-      />
-      <path
-        d="M22.6699 14.8269L22.6696 14.8262C21.8732 13.3622 19.8748 12.9039 18.2144 13.8049C16.5546 14.707 15.8507 16.6321 16.645 18.0959C16.9346 18.6299 17.3843 19.0304 17.9193 19.2798C18.8518 19.7147 20.0439 19.6911 21.1004 19.1182C22.7601 18.2161 23.4642 16.2911 22.6699 14.8269ZM20.4293 17.8818C19.4491 18.4133 18.3064 18.2084 17.8816 17.4249C17.4569 16.6424 17.9076 15.5729 18.8859 15.0411C19.5359 14.6884 20.2581 14.6603 20.7949 14.9107C21.0664 15.0373 21.2906 15.2351 21.4337 15.4984C21.8576 16.2812 21.4069 17.3503 20.4293 17.8818Z"
-        fill="#FEF9F9"
-      />
-    </g>
-    <defs>
-      <clipPath id="clip0_40956_131">
-        <rect width="24" height="24" fill="white" />
-      </clipPath>
-    </defs>
-  </svg>
-);}
-
-export default Paw;
-
