+++ src/components/Logo/Logo.styled.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
- 
+
 const MainLogo = styled.img`
 margin:0;
 @media screen and(min - width: 768px) {
