+++ src/components/Logout/Logout.jsx
@@ -3,3 +3,7 @@
 // const Logout = () => { return <LogoutButton></LogoutButton>}
 
 // export default Logout;
+
+// const Logout = () => { return <Button text='Logout'/> }
+
+// export default Logout;
