--- src/components/MobileMenu/MobileMenu.js
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-
-
-const MenuM = styled.div`
-${'' /* 
-display: block;
-  @media screen and (min-width: 768px) {
-display: block;
-  }
-
-  @media screen and (min-width: 1280px) { */}
-    ${'' /* display: none; */}
-    ${'' /* width: 500px;
-    height: 200px;
-  } */}
-`;
-
-export default MenuM;
