+++ src/components/MobileMenu/MobileMenu.jsx
@@ -1,13 +1,37 @@
-import MenuM from './MobileMenu';
-// import Logo from '../Logo/Logo';
-// import Nav from '../Nav/Nav';
-// import MenuLink from '../Header/MenuLink/MenuLink';
+import { Link, CloseButton, ButtonText, TopMenu } from './MobileMenu.styled';
+import MobileMenu from './MobileMenu';
+import Logo from '../Logo/Logo';
+import Nav from '../Nav/Nav';
+import Icon from 'utils/Icon/Icon';
+import Logout from 'components/Logout/Logout.styled';
 
-const MobileMenu = () => {
-  return (
-  <MenuM>
-    </MenuM>
-  );
-};
 
-export default MobileMenu;
+const Mobi = () => {
+    return (
+      <MobileMenu>
+        <TopMenu>
+        <div>
+          <Logo />
+        </div>
+        <div>
+          <Logout type="button">
+            <ButtonText color="logout" weight="bold" margin="8px">
+              Logout
+            </ButtonText>
+            <Icon name="logout" color="#FEF9F9" size={24} />
+          </Logout>
+          <CloseButton type="button">
+            <Icon name="cross" color="#FFC107" size={24} />
+          </CloseButton>
+        </div>
+        </TopMenu>
+        <Nav>
+          <Link to="/news">News</Link>
+          <Link to="/notices">Find Pet</Link>
+          <Link to="/friends">Our Friends</Link>
+        </Nav>
+      </MobileMenu>
+    );
+}
+
+export default Mobi;
