--- src/components/Nav/NavContainer/NavContainer.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const Container = styled.div`
-display: flex;
-justify-content: `;
-
-export default Container;
