--- src/components/Nav/NavContainer/NavContainer.jsx
@@ -1,5 +1,0 @@
-import Container from './NavContainer';
-
-const NavContainer = () => { return <Container></Container>; }
-
-export default NavContainer;
