--- src/components/UserNav/UserNavContainer/UserNavContainer.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const UserNavContainerM = styled.div`
-  display: flex;
-
-`;
-
-export default UserNavContainerM;
