--- src/components/UserNav/UserNavContainer/UserNavContainer.jsx
@@ -1,7 +1,0 @@
-import UserNavContainterM from './UserNavContainer';
-
-const UserNavContainer = () => {
-  return <UserNavContainterM></UserNavContainterM>;
-};
-
-export default UserNavContainer;
